Update WorldMania_English.lang
Fixed a couple of misspellings and clarity issues such as, “Player not founded”, changed to “Player not found!”, (Which I assume is what was supposed to be said) and changed “doesn’t exists”, to “doesn’t exist”. I really like the look of this project and I would like to be able to help with Japanese language support (as I also know a bit of Japanese). Hope you find this helpful!